pyproject.toml
[tool.poetry]
name = "scan-pdf"
version = "0.4.1"
description = "Tools for using scanners with document feeder"
authors = ["Andreas Würl <andi@tryb.de>"]
license = "Apache-2.0"
readme = "README.rst"

[tool.poetry.scripts]
scan-pdf = 'scan_pdf.console:run'

[tool.poetry.dependencies]
python = "^3.8"
pypdf = "^4.2.0"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
mock = "^5.1.0"
assertpy = "^1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
